Commit message should be:
"Refactor ImagePane.fxml and implement blur effect

This commit refactors the imports in ImagePane.fxml, changing import statements to a generic import for layout and text classes to reduce redundancy. Also, it renames views to more descriptive names, adjusting padding and adding 'Blur' button to UI.

The main purpose of this commit is to implement a blur effect on images using QuadTree. Here, each leaf node corresponds to the region of the image with the single solid color. The blur effect is achieved by visualizing the leaves at different levels of the QuadTree, where each level corresponds to a different degree of image detail.

A new animation feature is also added such that the image, starting from lowest level of detail (completely blurred), gradually increase to the highest level of detail (original image). This is implemented by storing a list of images corresponding to different levels of detail during encoding, then using a timeline to cycle through these images."