Added animated rendering of region tree
Changes were made to ImagePaneController to improve the visualization of the RegionQuadTree. This was done by implementing squares animation. When constructing the tree, instead of directly adding rectangles to the pane, they are now stored in a Stack. This new timeline 'timeline2' then pops each rectangle and adds it to the pane after a certain duration, giving an animated effect. An 'isAnimated' flag was also added to control the behavior of 'drawSquares' and 'showTreeSquares' methods based on whether we want the animation or not. With these modifications, the user can better understand how the RegionQuadTree is constructed.