Refactor QuadTree data structure from HashSet to List
The primary change in this commit is the refactoring of data structure being used in QuadTree from HashSet to List. This change makes the data structure more predictable as now the data order is preserved. The HashSet was replaced as it caused unnecessary complexity due to its unordered nature.

Additional modifications were made to improve the efficiency and readability, including the addition of capacity and clearList parameters to PointQuadTree constructors to control the maximum number of points allowed in a leaf node and whether the list of elements was cleared after partitioning respectively.

In the PointQuadTree query method, an additional check that the query rectangle intersects with the QuadTree square has been added to improve performance. This is because it stops unnecessary further calculations when the QuadTree square and query rectangle do not overlap at all.

Finally, the KDTreeEfficient's query method also saw the same replacement of HashSet with List, and modifications needed for the above changes to work were carried across all affected methods and classes.